fix aany

+++ Website_NoRulesShop/src/main/java/com/poly/website_norulesshop/Repository/BrandImageRepository.java
@@ -1,10 +1,11 @@
 package com.poly.website_norulesshop.Repository;
 
-import com.poly.website_norulesshop.entity.BrandImage;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import com.poly.website_norulesshop.entity.BrandImage;
+
 @Repository
-public interface BrandImageRepository extends JpaRepository<BrandImage, Long> {
+public interface BrandImageRepository extends JpaRepository<BrandImage, Integer> {
     // Các phương thức đặc thù cho bảng BrandImage (nếu cần)
 }
