+++ Website_NoRulesShop/src/main/java/com/poly/website_norulesshop/Repository/CategoryLevel1DetailRepository.java
@@ -1,10 +1,11 @@
 package com.poly.website_norulesshop.Repository;
 
-import com.poly.website_norulesshop.entity.CategoryLevel1Detail;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import com.poly.website_norulesshop.entity.CategoryLevel1Detail;
+
 @Repository
-public interface CategoryLevel1DetailRepository extends JpaRepository<CategoryLevel1Detail, Long> {
+public interface CategoryLevel1DetailRepository extends JpaRepository<CategoryLevel1Detail, Integer> {
     // Các phương thức đặc thù cho bảng CategoryLevel1Detail (nếu cần)
 }
