+++ Website_NoRulesShop/src/main/java/com/poly/website_norulesshop/Repository/CategoryLevel2DetailRepository.java
@@ -1,10 +1,11 @@
 package com.poly.website_norulesshop.Repository;
 
-import com.poly.website_norulesshop.entity.CategoryLevel2Detail;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import com.poly.website_norulesshop.entity.CategoryLevel2Detail;
+
 @Repository
-public interface CategoryLevel2DetailRepository extends JpaRepository<CategoryLevel2Detail, Long> {
+public interface CategoryLevel2DetailRepository extends JpaRepository<CategoryLevel2Detail, Integer> {
     // Các phương thức đặc thù cho bảng CategoryLevel2Detail (nếu cần)
 }
