+++ Website_NoRulesShop/src/main/java/com/poly/website_norulesshop/Repository/FeedbackImageRepository.java
@@ -1,10 +1,11 @@
 package com.poly.website_norulesshop.Repository;
 
-import com.poly.website_norulesshop.entity.FeedbackImage;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import com.poly.website_norulesshop.entity.FeedbackImage;
+
 @Repository
-public interface FeedbackImageRepository extends JpaRepository<FeedbackImage, Long> {
+public interface FeedbackImageRepository extends JpaRepository<FeedbackImage, Integer> {
     // Các phương thức đặc thù cho bảng FeedbackImage (nếu cần)
 }
