+++ Website_NoRulesShop/src/main/java/com/poly/website_norulesshop/Repository/PaymentMethodRepository.java
@@ -1,10 +1,11 @@
 package com.poly.website_norulesshop.Repository;
 
-import com.poly.website_norulesshop.entity.PaymentMethod;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import com.poly.website_norulesshop.entity.PaymentMethod;
+
 @Repository
-public interface PaymentMethodRepository extends JpaRepository<PaymentMethod, Long> {
+public interface PaymentMethodRepository extends JpaRepository<PaymentMethod, Integer> {
     // Các phương thức đặc thù cho bảng PaymentMethod (nếu cần)
 }
