+++ Website_NoRulesShop/src/main/java/com/poly/website_norulesshop/Repository/PointsRepository.java
@@ -1,10 +1,11 @@
 package com.poly.website_norulesshop.Repository;
 
-import com.poly.website_norulesshop.entity.Points;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import com.poly.website_norulesshop.entity.Points;
+
 @Repository
-public interface PointsRepository extends JpaRepository<Points, Long> {
+public interface PointsRepository extends JpaRepository<Points, Integer> {
     // Các phương thức đặc thù cho bảng Points (nếu cần)
 }
