+++ Website_NoRulesShop/src/main/java/com/poly/website_norulesshop/Repository/VoucherRepository.java
@@ -1,10 +1,11 @@
 package com.poly.website_norulesshop.Repository;
 
-import com.poly.website_norulesshop.entity.Voucher;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import com.poly.website_norulesshop.entity.Voucher;
+
 @Repository
-public interface VoucherRepository extends JpaRepository<Voucher, Long> {
+public interface VoucherRepository extends JpaRepository<Voucher, Integer> {
     // Các phương thức đặc thù cho bảng Voucher (nếu cần)
 }
