+++ Website_NoRulesShop/src/main/java/com/poly/website_norulesshop/service/BrandImageService.java
@@ -1,16 +1,17 @@
 package com.poly.website_norulesshop.service;
 
-import com.poly.website_norulesshop.entity.BrandImage;
+import java.util.List;
+
 import org.springframework.stereotype.Service;
 
-import java.util.List;
+import com.poly.website_norulesshop.entity.BrandImage;
 @Service
 public interface BrandImageService {
     BrandImage saveBrandImage(BrandImage brandImage);
 
-    BrandImage getBrandImageById(Long id);
+    BrandImage getBrandImageById(Integer id);
 
     List<BrandImage> getAllBrandImages();
 
-    void deleteBrandImage(Long id);
+    void deleteBrandImage(Integer id);
 }
