+++ Website_NoRulesShop/src/main/java/com/poly/website_norulesshop/service/CategoryLevel1DetailService.java
@@ -1,16 +1,17 @@
 package com.poly.website_norulesshop.service;
 
-import com.poly.website_norulesshop.entity.CategoryLevel1Detail;
+import java.util.List;
+
 import org.springframework.stereotype.Service;
 
-import java.util.List;
+import com.poly.website_norulesshop.entity.CategoryLevel1Detail;
 @Service
 public interface CategoryLevel1DetailService {
     CategoryLevel1Detail saveCategoryLevel1Detail(CategoryLevel1Detail categoryLevel1Detail);
 
-    CategoryLevel1Detail getCategoryLevel1DetailById(Long id);
+    CategoryLevel1Detail getCategoryLevel1DetailById(Integer id);
 
     List<CategoryLevel1Detail> getAllCategoryLevel1Details();
 
-    void deleteCategoryLevel1Detail(Long id);
+    void deleteCategoryLevel1Detail(Integer id);
 }
