+++ Website_NoRulesShop/src/main/java/com/poly/website_norulesshop/service/CategoryLevel1Service.java
@@ -1,16 +1,17 @@
 package com.poly.website_norulesshop.service;
 
-import com.poly.website_norulesshop.entity.CategoryLevel1;
+import java.util.List;
+
 import org.springframework.stereotype.Service;
 
-import java.util.List;
+import com.poly.website_norulesshop.entity.CategoryLevel1;
 @Service
 public interface CategoryLevel1Service {
     CategoryLevel1 saveCategoryLevel1(CategoryLevel1 categoryLevel1);
 
-    CategoryLevel1 getCategoryLevel1ById(Long id);
+    CategoryLevel1 getCategoryLevel1ById(Integer id);
 
     List<CategoryLevel1> getAllCategoryLevel1s();
 
-    void deleteCategoryLevel1(Long id);
+    void deleteCategoryLevel1(Integer id);
 }
