+++ Website_NoRulesShop/src/main/java/com/poly/website_norulesshop/service/CategoryLevel2DetailService.java
@@ -1,16 +1,17 @@
 package com.poly.website_norulesshop.service;
 
-import com.poly.website_norulesshop.entity.CategoryLevel2Detail;
+import java.util.List;
+
 import org.springframework.stereotype.Service;
 
-import java.util.List;
+import com.poly.website_norulesshop.entity.CategoryLevel2Detail;
 @Service
 public interface CategoryLevel2DetailService {
     CategoryLevel2Detail saveCategoryLevel2Detail(CategoryLevel2Detail categoryLevel2Detail);
 
-    CategoryLevel2Detail getCategoryLevel2DetailById(Long id);
+    CategoryLevel2Detail getCategoryLevel2DetailById(Integer id);
 
     List<CategoryLevel2Detail> getAllCategoryLevel2Details();
 
-    void deleteCategoryLevel2Detail(Long id);
+    void deleteCategoryLevel2Detail(Integer id);
 }
