+++ Website_NoRulesShop/src/main/java/com/poly/website_norulesshop/service/CategoryLevel2Service.java
@@ -1,16 +1,17 @@
 package com.poly.website_norulesshop.service;
 
-import com.poly.website_norulesshop.entity.CategoryLevel2;
+import java.util.List;
+
 import org.springframework.stereotype.Service;
 
-import java.util.List;
+import com.poly.website_norulesshop.entity.CategoryLevel2;
 @Service
 public interface CategoryLevel2Service {
     CategoryLevel2 saveCategoryLevel2(CategoryLevel2 categoryLevel2);
 
-    CategoryLevel2 getCategoryLevel2ById(Long id);
+    CategoryLevel2 getCategoryLevel2ById(Integer id);
 
     List<CategoryLevel2> getAllCategoryLevel2s();
 
-    void deleteCategoryLevel2(Long id);
+    void deleteCategoryLevel2(Integer id);
 }
