+++ Website_NoRulesShop/src/main/java/com/poly/website_norulesshop/service/Impl/FeedbackServiceImpl.java
@@ -1,12 +1,13 @@
 package com.poly.website_norulesshop.service.Impl;
 
-import com.poly.website_norulesshop.Repository.FeedbackRepository;
-import com.poly.website_norulesshop.entity.Feedback;
-import com.poly.website_norulesshop.service.FeedbackService;
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
+import com.poly.website_norulesshop.Repository.FeedbackRepository;
+import com.poly.website_norulesshop.entity.Feedback;
+import com.poly.website_norulesshop.service.FeedbackService;
 
 @Service
 public class FeedbackServiceImpl implements FeedbackService {
@@ -24,7 +25,7 @@ public class FeedbackServiceImpl implements FeedbackService {
     }
 
     @Override
-    public Feedback getFeedbackById(Long id) {
+    public Feedback getFeedbackById(Integer id) {
         return feedbackRepository.findById(id).orElse(null);
     }
 
@@ -34,7 +35,7 @@ public class FeedbackServiceImpl implements FeedbackService {
     }
 
     @Override
-    public void deleteFeedback(Long id) {
+    public void deleteFeedback(Integer id) {
         feedbackRepository.deleteById(id);
     }
 }
