+++ Website_NoRulesShop/src/main/java/com/poly/website_norulesshop/service/PaymentMethodService.java
@@ -1,16 +1,17 @@
 package com.poly.website_norulesshop.service;
 
-import com.poly.website_norulesshop.entity.PaymentMethod;
+import java.util.List;
+
 import org.springframework.stereotype.Service;
 
-import java.util.List;
+import com.poly.website_norulesshop.entity.PaymentMethod;
 @Service
 public interface PaymentMethodService {
     PaymentMethod savePaymentMethod(PaymentMethod paymentMethod);
 
-    PaymentMethod getPaymentMethodById(Long id);
+    PaymentMethod getPaymentMethodById(Integer id);
 
     List<PaymentMethod> getAllPaymentMethods();
 
-    void deletePaymentMethod(Long id);
+    void deletePaymentMethod(Integer id);
 }
